Allow parsing token accounts as json
Add an optional DB trait that offers a specialized GPA method allowing easier json serialization of token account.

The new trait offers a method that returns token account data with a mint pubkey and a hash of deserialized mint accounts for the queried token accounts that are necessary to correctly serialize token accounts in json format.